minor tweaks to cmd to remove version references
