+++ cmd/root.go
@@ -8,13 +8,8 @@ import (
 	inertiacmd "github.com/ubclaunchpad/inertia/cmd/cmd"
 	configcmd "github.com/ubclaunchpad/inertia/cmd/config"
 	hostcmd "github.com/ubclaunchpad/inertia/cmd/host"
-	"github.com/ubclaunchpad/inertia/cmd/inpututil"
-	"github.com/ubclaunchpad/inertia/cmd/printutil"
 	provisioncmd "github.com/ubclaunchpad/inertia/cmd/provision"
 	remotecmd "github.com/ubclaunchpad/inertia/cmd/remote"
-
-	"github.com/ubclaunchpad/inertia/common"
-	"github.com/ubclaunchpad/inertia/local"
 )
 
 func getVersion(version string) string {
@@ -70,64 +65,4 @@ Issue tracker: https://github.com/ubclaunchpad/inertia/issues`,
 	attachContribPlugins(root)
 
 	return root
-}
-
-func attachInitCmd(inertia *inertiacmd.Cmd) {
-	const flagVersion = "version"
-	var init = &cobra.Command{
-		Use:   "init",
-		Short: "Initialize an Inertia project in this repository",
-		Long: `Initializes an Inertia project in this GitHub repository.
-		There must be a local git repository in order for initialization
-		to succeed.`,
-		Run: func(cmd *cobra.Command, args []string) {
-			version := cmd.Parent().Version
-			givenVersion, err := cmd.Flags().GetString(flagVersion)
-			if err != nil {
-				printutil.Fatal(err)
-			}
-			if givenVersion != version {
-				version = givenVersion
-			}
-
-			// Determine best build type for project
-			var buildType string
-			var buildFilePath string
-			cwd, err := os.Getwd()
-			if err != nil {
-				printutil.Fatal(err)
-			}
-			// docker-compose projects will usually have Dockerfiles,
-			// so check for that first, then check for Dockerfile
-			if common.CheckForDockerCompose(cwd) {
-				println("docker-compose project detected")
-				buildType = "docker-compose"
-				buildFilePath = "docker-compose.yml"
-			} else if common.CheckForDockerfile(cwd) {
-				println("Dockerfile project detected")
-				buildType = "dockerfile"
-				buildFilePath = "Dockerfile"
-			} else {
-				println("No build file detected")
-				buildType, buildFilePath, err = inpututil.AddProjectWalkthrough(os.Stdin)
-				if err != nil {
-					printutil.Fatal(err)
-				}
-			}
-
-			// Hello world config file!
-			err = local.InitializeInertiaProject(inertia.ConfigPath, inertia.Version, buildType, buildFilePath)
-			if err != nil {
-				printutil.Fatal(err)
-			}
-			println("An inertia.toml configuration file has been created to store")
-			println("Inertia configuration. It is recommended that you DO NOT commit")
-			println("this file in source control since it will be used to store")
-			println("sensitive information.")
-			println("\nYou can now use 'inertia remote add' to connect your remote")
-			println("VPS instance.")
-		},
-	}
-	init.Flags().String(flagVersion, inertia.Version, "specify Inertia daemon version to use")
-	inertia.AddCommand(init)
 }
