+++ contrib/inertia-docgen/main.go
@@ -2,7 +2,9 @@ package main
 
 import (
 	"fmt"
+	"io/ioutil"
 	"os"
+	"path/filepath"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/cobra/doc"
@@ -16,7 +18,17 @@ import (
 )
 
 // Version denotes the version of the binary
-var Version string
+var (
+	Version string
+
+	mdReadmeTemplate = `# Inertia Command Reference
+
+Click [here](/inertia.md) for the Inertia CLI command reference. It is generated
+automatically using ` + "`inertia-docgen`." + `
+
+For a more general usage guide, refer to the [Inertia Usage Guide](https://inertia.ubclaunchpad.com).
+`
+)
 
 func main() {
 	var root = cmd.NewInertiaCmd(Version)
@@ -32,9 +44,10 @@ func getDocgenCmd(root *inertiacmd.Cmd) *cobra.Command {
 		flagFormat = "format"
 	)
 	var docs = &cobra.Command{
-		Use:    "inertia-docgen",
-		Hidden: true,
-		Short:  "Generate command reference for the Inertia CLI.",
+		Use:     "inertia-docgen",
+		Hidden:  true,
+		Version: Version,
+		Short:   "Generate command reference for the Inertia CLI.",
 		Run: func(cmd *cobra.Command, args []string) {
 			var output, _ = cmd.Flags().GetString(flagOutput)
 			var format, _ = cmd.Flags().GetString(flagFormat)
@@ -50,21 +63,22 @@ func getDocgenCmd(root *inertiacmd.Cmd) *cobra.Command {
 			os.MkdirAll(output, os.ModePerm)
 
 			// gen docs
-			var err error
 			switch format {
 			case "man":
-				err = doc.GenManTree(root.Command, &doc.GenManHeader{
+				if err := doc.GenManTree(root.Command, &doc.GenManHeader{
 					Title: "Inertia CLI",
 					Source: fmt.Sprintf(
 						"Generated by Inertia %s (https://github.com/ubclaunchpad/inertia)",
 						root.Version),
 					Manual: "https://inertia.ubclaunchpad.com",
-				}, output)
+				}, output); err != nil {
+					printutil.Fatal(err.Error())
+				}
 			default:
-				err = doc.GenMarkdownTree(root.Command, output)
-			}
-			if err != nil {
-				printutil.Fatal(err.Error())
+				if err := doc.GenMarkdownTree(root.Command, output); err != nil {
+					printutil.Fatal(err.Error())
+				}
+				ioutil.WriteFile(filepath.Join(output, "README.md"), []byte(mdReadmeTemplate), os.ModePerm)
 			}
 
 			fmt.Printf("%s documentation generated in %s", format, output)
