+++ docs/cli/README.md
@@ -3,4 +3,4 @@
 Click [here](/inertia.md) for the Inertia CLI command reference. It is generated
 automatically using `inertia-docgen`.
 
-For a more general usage guide, refer to the [Inertia Usage Guide](https://inertia.ubclaunchpad.com)
+For a more general usage guide, refer to the [Inertia Usage Guide](https://inertia.ubclaunchpad.com).
