--- docs/cli/inertia_x.md
@@ -1,26 +1,0 @@
-## inertia x
-
-Experimental Inertia commands and features
-
-### Synopsis
-
-Experimental Inertia commands, features, and utilities - some are intended
-for development use only, and others are simply in-development features.
-
-### Options
-
-```
-  -h, --help   help for x
-```
-
-### Options inherited from parent commands
-
-```
-      --config string   specify relative path to Inertia configuration (default "inertia.toml")
-```
-
-### SEE ALSO
-
-* [inertia](inertia.md)	 - Effortless, self-hosted continuous deployment for small teams and projects
-* [inertia x docs](inertia_x_docs.md)	 - List currently configured remotes
-
