--- docs/cli/inertia_x_docs.md
@@ -1,30 +1,0 @@
-## inertia x docs
-
-List currently configured remotes
-
-### Synopsis
-
-Lists all currently configured remotes.
-
-```
-inertia x docs [flags]
-```
-
-### Options
-
-```
-  -f, --format string   format to generate (md|man) (default "md")
-  -h, --help            help for docs
-  -o, --output string   output file path (default "./inertia_cli_reference")
-```
-
-### Options inherited from parent commands
-
-```
-      --config string   specify relative path to Inertia configuration (default "inertia.toml")
-```
-
-### SEE ALSO
-
-* [inertia x](inertia_x.md)	 - Experimental Inertia commands and features
-
